# dev/doc/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Formula 1 API",
    "description": "API for Formula 1 World Championship data",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
...
  "paths": {
...
    "/locations/{locationId}": {
      "get": {
        "tags": [
          "location"
        ],
        "summary": "Get a location by ID",
        "operationId": "getLocationById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "description": "ID of the location",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          },
          "404": {
            "description": "Location not found"
          },
          "500": {
            "description": "An error occurred on the server (Internal Server Error)"
          }
        }
      },
      "put": {
        "tags": [
          "location"
        ],
        "summary": "Update a location by ID",
        "operationId": "updateLocationById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "description": "ID of the location",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Location updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          },
          "404": {
            "description": "Location not found"
          },
          "500": {
            "description": "An error occurred on the server (Internal Server Error)"
          }
        }
      },
      "delete": {
        "tags": [
          "location"
        ],
        "summary": "Delete a location by ID",
        "operationId": "deleteLocationById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "description": "ID of the location",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Location deleted successfully"
          },
          "404": {
            "description": "Location not found"
          }
        }
      }
    },
...
    "/countries/{countryId}": {
      "get": {
        "tags": [
          "country"
        ],
        "summary": "Get a country by ID",
        "operationId": "getCountryById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "countryId",
            "in": "path",
            "description": "ID of the country",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              }
            }
          },
          "404": {
            "description": "Country not found"
          },
          "500": {
            "description": "An error occurred on the server (Internal Server Error)"
          }
        }
      },
      "put": {
        "tags": [
          "country"
        ],
        "summary": "Update a country by ID",
        "operationId": "updateCountryById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "countryId",
            "in": "path",
            "description": "ID of the country",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CountryInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Country updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              }
            }
          },
          "404": {
            "description": "Country not found"
          },
          "500": {
            "description": "An error occurred on the server (Internal Server Error)"
          }
        }
      },
      "delete": {
        "tags": [
          "country"
        ],
        "summary": "Delete a country by ID",
        "operationId": "deleteCountryById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "countryId",
            "in": "path",
            "description": "ID of the country",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Country deleted successfully"
          },
          "404": {
            "description": "Country not found"
          }
        }
      }
    },
...
    "/users/{userId}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get info about User",
        "operationId": "getUserById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "An error occurred on the server (Internal Server Error)"
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Delete a specific User",
        "operationId": "deleteUserById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "An error occurred on the server (Internal Server Error)"
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update a specific User",
        "operationId": "updateUserById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "An error occurred on the server (Internal Server Error)"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Location": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "location": {
            "type": "string"
          }
        }
      },
      "LocationInput": {
        "type": "object",
        "properties": {
          "location": {
            "type": "string"
          }
        }
      },
      "Country": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        }
      },
      "CountryInput": {
        "type": "object",
        "properties": {
          "country": {
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
      "UserInput": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}